fix(nvim): Remove mini-diff extra.
Mini.diff overlaps with gitsigns in a lot of functionality, which is why
Folke disables gitsigns when you enable the Mini.diff extra. However
TONS of plugins integrate with gitsigns while Mini.diff is too new for
said integrations. As much as I love the diff overlay functionality,
it's not worth trying to use both plugins and configuring them to play
nice.