fix: Rewrite symlink creation to avoid associative array which is only available in Bash 4.x or greater.
Symlink creation is intended to run on a fresh macbook or linux install,
but Apple does not bother to ship updated bash anymore. We need to write
more old school bash code for this to avoid bash version issues. We
could use zsh but then it won't run on any linux distro by default.